+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- [**breaking**] Optionally allow starting threads with `Timeline::send_reply`.
+  ([4819](https://github.com/matrix-org/matrix-rust-sdk/pull/4819))
+
 ### Refactor
 
 - [**breaking**] Reactions on a given timeline item have been moved from